fix: Primary btn disables on error message, Route to patient form if already exist, Rename Sandbox_config to abdm_config, clean up utils
